requirements.txt
flask 
google-generativeai
gunicorn
markdown
wikipedia